# openapi_tech_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Система бронирования мест в кафе",
    "version": "0.0.3"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1/",
      "description": "Локальный сервер"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Получение списка пользователей",
        "description": "Возвращает информацию о всех пользователях. Только для администраторов или менеджеров",
        "operationId": "get_all_user_users_get",
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserInfo"
                  },
                  "type": "array",
                  "title": "Информация о пользователях"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Регистрация нового пользователя",
        "description": "Создает нового пользователя с указанными данными. <br> Регистрировать пользователя может или не авторизированный пользователь или менеджер или администратор.<br><br><b>Обязательные поля:</b> <ul><li>username</li><li>password</li><li>email или phone</li></ul>",
        "operationId": "create_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomError": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "CustomError"
      },
...
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "tg_id": {
            "type": "string",
            "title": "Tg Id"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "tg_id": {
            "type": "string",
            "title": "Tg Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "UserInfo"
      },
      "UserRole": {
        "type": "string",
        "oneOf": [
          {
            "title": "admin",
            "const": "ADMIN",
            "description": "Администратор"
          },
          {
            "title": "manager",
            "const": "MANAGER",
            "description": "Менеджер"
          },
          {
            "title": "user",
            "const": "USER",
            "description": "Пользователь"
          }
        ],
        "title": "UserRole"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}